kvstorage: assert each engine only access its spans
The commit adds engine assertions to assert that Log
engine only accesses log engine spans, and state engine
only accesses state engine spans.
