+++ pkg/kv/kvserver/kvstorage/BUILD.bazel
@@ -21,11 +21,13 @@ go_library(
         "//pkg/kv/kvserver/kvserverpb",
         "//pkg/kv/kvserver/logstore",
         "//pkg/kv/kvserver/rditer",
+        "//pkg/kv/kvserver/spanset",
         "//pkg/raft/raftpb",
         "//pkg/roachpb",
         "//pkg/storage",
         "//pkg/storage/enginepb",
         "//pkg/storage/fs",
+        "//pkg/util",
         "//pkg/util/buildutil",
         "//pkg/util/hlc",
         "//pkg/util/iterutil",
@@ -45,6 +47,7 @@ go_test(
         "init_test.go",
         "initial_test.go",
         "stateloader_test.go",
+        "storage_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":kvstorage"],
@@ -56,6 +59,7 @@ go_test(
         "//pkg/kv/kvserver/kvserverpb",
         "//pkg/kv/kvserver/logstore",
         "//pkg/kv/kvserver/print",
+        "//pkg/kv/kvserver/spanset",
         "//pkg/raft/raftpb",
         "//pkg/roachpb",
         "//pkg/settings/cluster",
