+++ pkg/kv/kvserver/replica_read.go
@@ -94,7 +94,6 @@ func (r *Replica) executeReadOnlyBatch(
 	}
 	if util.RaceEnabled {
 		spans := g.LatchSpans()
-		spans.AddForbiddenMatcher(validateIsStateEngineSpan)
 		rw = spanset.NewReadWriterAt(rw, spans, ba.Timestamp)
 	}
 	defer rw.Close()
